+++ extensions/omni.isaac.orbit_quadruped_cpg/docs/CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.1] - 2024-01-29
+
+### Added
+
+- Created `get_all_leg_joints` method in `QuadrupedIK` class.
+
+### Fixed
+
+- Fixed `GO2_IK_CFG` lengths.
+- Fixed `get_leg_joints` method of `QuadrupedIK` return shape.
+
+
 ## [0.1.0] - 2024-01-28
 
 ### Added